# server/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DealForge API",
    "description": "AI-powered meeting intelligence and sales CRM API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "API server" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Read-only API key (generated in Settings → API Access). Available on Pro."
      }
    },
    "schemas": {
...
      "MeetingWithAnalysis": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "startTime": { "type": "string" },
          "endTime": { "type": "string" },
          "duration": { "type": "number" },
          "status": { "type": "string" },
          "summary": { "type": "string", "nullable": true },
          "actionItems": { "type": "array", "items": { "type": "string" } },
          "leadScore": { "type": "number", "nullable": true },
          "modelUsed": { "type": "string", "nullable": true },
          "analyzedAt": { "type": "string", "nullable": true }
        }
      }
    }
  },
  "paths": {
...
    "/referrals/claim": {
      "post": {
        "summary": "Claim a referral code",
        "description": "Records that the authenticated user was referred by a code. Free users get +1 analyzed meeting/month for 3 months; paid users get 1 free month credit.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code"],
                "properties": {
                  "code": { "type": "string", "example": "DF-ABCD2345" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Claim result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "claimStatus": { "type": "string", "enum": ["claimed", "already_claimed", "self_referral", "invalid_code", "limit_reached"] },
                    "benefit": { "type": "string", "enum": ["meeting_bonus", "free_month"] },
                    "code": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/referrals/status": {
      "get": {
        "summary": "Get referral status",
        "description": "Returns the user's referral code, claimed codes, active benefits, and how many people used their code.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Referral status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "code": { "type": "string" },
                    "bonusMeetings": { "type": "integer" },
                    "freeMonthsCredit": { "type": "integer" },
                    "referralsMade": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "uid": { "type": "string" },
                          "claimedAt": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/v1/meetings": {
      "get": {
        "summary": "List meetings with summaries and action items",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "Meetings with joined analysis data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/MeetingWithAnalysis" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Missing or invalid API key" }
        }
      }
    },
    "/v1/meetings/{id}": {
      "get": {
        "summary": "Get a single meeting with its analysis",
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Meeting detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "$ref": "#/components/schemas/MeetingWithAnalysis" }
                  }
                }
              }
            }
          },
          "401": { "description": "Missing or invalid API key" },
          "404": { "description": "Meeting not found" }
        }
      }
    },
    "/v1/leads": {
      "get": {
        "summary": "List leads",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "Leads",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/SyncPayload/properties/leads/items" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Missing or invalid API key" }
        }
      }
    },
    "/v1/deals": {
      "get": {
        "summary": "List deals",
        "security": [{ "ApiKeyAuth": [] }],
        "responses": {
          "200": {
            "description": "Deals",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/SyncPayload/properties/deals/items" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Missing or invalid API key" }
        }
      }
    }
  }
}